software/requirements.txt
Pillow>=10.0.0
pyyaml>=6.0
asyncssh>=2.14.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
webrtcvad>=2.0.10
sounddevice>=0.4.6
numpy>=1.24.0
